Refactor gpu_timer into rl folder
