+++ torchtitan/experiments/rl/metrics.py
@@ -8,7 +8,45 @@
 Shared metrics utilities for RL training loops.
 """
 
+import time
+from contextlib import contextmanager
 from dataclasses import dataclass
+from typing import Generator
+
+from torchtitan.tools.utils import device_module
+
+
+@contextmanager
+def gpu_timer(
+    sync: bool = True, enabled: bool = True
+) -> Generator[dict[str, float], None, None]:
+    """Context manager that optionally synchronizes the device and measures wall-clock time.
+
+    Yields a dict; on exit it is populated with an ``elapsed_s`` key.
+
+    Args:
+        sync: If True, synchronizes the device before and after timing.
+        enabled: If False, skips all timing/sync overhead and yields an empty dict.
+
+    Usage::
+
+        with gpu_timer(sync=cuda_sync_for_metrics, enabled=enable_profiling) as t:
+            do_work()
+        print(t.get("elapsed_s", 0.0))
+    """
+    result: dict[str, float] = {}
+    if not enabled:
+        yield result
+        return
+    if sync:
+        device_module.synchronize()
+    t0 = time.perf_counter()
+    try:
+        yield result
+    finally:
+        if sync:
+            device_module.synchronize()
+        result["elapsed_s"] = time.perf_counter() - t0
 
 
 @dataclass
